+++ src/gulcalc/getrands.h
@@ -52,13 +52,13 @@ class getRands {
 private:
 	rd_option rndopt_;
 	int rand_vec_size_;
-	float *buf_;
+	double *buf_;
 	int base_offset_;
 	unsigned int buffersize_;
 	std::mt19937 gen_;
 	std::uniform_real_distribution<> dis_;
     //int _randsamplesize;
-    std::vector<float> rnd_;
+    std::vector<double> rnd_;
 	int rand_seed_;
 	void userandfile();
 public:
@@ -73,9 +73,9 @@ public:
         rnd_.resize(rand_vec_size_,-1);
     }
 
-	inline float nextrnd() { return (float) dis_(gen_); }	// used after seeding via group id and event_id
+	inline double nextrnd() { return (double) dis_(gen_); }	// used after seeding via group id and event_id
 
-	inline float rnd(unsigned int ridx)  {
+	inline double rnd(unsigned int ridx)  {
 		switch (rndopt_) {
 			case rd_option::userandomnumberfile:
 			{
@@ -85,9 +85,9 @@ public:
 			break;
 			case rd_option::usecachedvector:
 			{
-				float f = rnd_[ridx % rand_vec_size_];
+				double f = rnd_[ridx % rand_vec_size_];
 				if (f < 0) {
-					f = (float)dis_(gen_);
+					f = (double)dis_(gen_);
 					rnd_[ridx % rand_vec_size_] = f;
 				}
 				return f;
@@ -95,7 +95,7 @@ public:
 			break;
 			case rd_option::usehashedseed: 
 			{
-				return (float)dis_(gen_);
+				return (double)dis_(gen_);
 			}
 			break;
 		}
